# Optimize `BalancedMerkleTree::compute_root_only()` using SIMD