Similarly to https://github.com/nazar-pc/abundance/pull/345, massive wins here again both for Merkle Tree root computation and archiving (which uses it) benchmarks:
```
Before:
65536/balanced/compute-root-only
                        time:   [4.0835 ms 4.0836 ms 4.0841 ms]
After:
65536/balanced/compute-root-only
                        time:   [555.84 µs 556.90 µs 557.80 µs]

Before:
segment-archiving-whole-segment
                        time:   [1.8868 s 1.8892 s 1.8914 s]
After:
segment-archiving-whole-segment
                        time:   [950.64 ms 957.24 ms 962.52 ms]
```

This is all on a single core. And archiving can be optimized further by making erasure coding more efficient and when BLAKE3 APIs are improved. Not to mention potential GPU acceleration.

Above benchmark is single-threaded, for context, single-threaded Subspace:
```
segment-archiving-whole-segment
                        time:   [39.370 s 39.370 s 39.370 s]
```

:exploding_head: